backend/VulnDefend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp_socks==0.9.0
aiosignal==1.3.1
alembic==1.13.1
altair==4.2.2
annotated-types==0.6.0
anyio==4.6.2.post1
asgiref==3.8.1
Aspose.Slides==24.7.0
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==24.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
bio==1.7.1
biopython==1.83
biothings-client==0.3.1
bleach==6.1.0
blinker==1.7.0
boto3==1.34.131
botocore==1.34.131
bs4==0.0.2
build==1.2.1
cachetools==5.5.0
cattrs==23.2.3
certifi==2024.8.30
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==1.0.12
click==8.1.7
cohere==5.5.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.20.0
dateparser==1.2.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
Django==5.1.2
django-cors-headers==4.6.0
django-crontab==0.7.1
django-filter==24.3
djangorestframework==3.15.2
dnspython==2.6.1
easyocr==1.7.1
elastic-transport==8.15.1
elasticsearch==8.15.1
entrypoints==0.4
executing==2.0.1
Faker==25.9.1
fastapi==0.115.9
fastavro==1.9.4
favicon==0.7.0
filelock==3.13.4
Flask==3.0.3
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.53.0
frozendict==2.4.4
frozenlist==1.5.0
fsspec==2024.5.0
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
googleapis-common-protos==1.63.0
gprofiler-official==1.0.0
gpt4all==2.7.0
greenlet==3.1.1
grpcio==1.64.1
grpcio-status==1.62.2
grpcio-tools==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htbuilder==0.6.2
htmldate==1.8.1
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.22.2
humanfriendly==10.0
humanize==4.9.0
hyperframe==6.0.1
idna==3.10
imageio==2.34.1
importlib-metadata==7.0.0
importlib_resources==6.4.0
InstructorEmbedding==1.0.1
ipython==8.23.0
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.11
jwt==1.3.1
kaggle==1.6.14
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.3.25
langchain-chroma==0.2.4
langchain-community==0.0.38
langchain-core==0.3.63
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.1.147
lazy_loader==0.4
llama-cloud==0.1.4
llama-index==0.10.68
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.68.post1
llama-index-embeddings-huggingface==0.2.3
llama-index-embeddings-instructor==0.1.3
llama-index-embeddings-langchain==0.1.2
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48
llama-index-llms-anyscale==0.1.4
llama-index-llms-fireworks==0.1.5
llama-index-llms-langchain==0.1.4
llama-index-llms-openai==0.1.31
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-index-storage-docstore-mongodb==0.1.3
llama-index-storage-index-store-firestore==0.2.0
llama-index-storage-index-store-mongodb==0.2.0
llama-index-storage-kvstore-firestore==0.2.1
llama-index-storage-kvstore-mongodb==0.1.2
llama-index-vector-stores-chroma==0.1.10
llama-index-vector-stores-mongodb==0.5.0
llama-index-vector-stores-qdrant==0.2.17
llama-parse==0.4.9
llamaindex-py-client==0.1.19
lxml==5.2.1
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.23.0
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
merkle-json==1.0.0
millify==0.1.1
minijinja==1.0.16
mmh3==4.1.0
monotonic==1.6
more-itertools==10.3.0
motor==3.4.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
munch==4.0.0
mygene==3.2.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.2.0
onnxruntime==1.17.1
open-clip-torch==2.24.0
openai==1.52.2
opencv-python-headless==4.9.0.80
openparse==0.6.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.18
ormsgpack==1.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
Pillow==10.0.0
pinecone-client==4.1.1
pinecone-plugin-interface==0.0.7
pip-tools==7.4.1
platformdirs==4.2.2
pooch==1.8.1
portalocker==2.8.2
posthog==3.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
propcache==0.2.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.17.2
PyJWT==2.8.0
pymdown-extensions==10.8.1
pymed==0.8.9
pymongo==4.7.2
PyMuPDF==1.24.11
pymupdf4llm==0.0.17
PyMuPDFb==1.24.1
pyparsing==3.2.0
pypdf==4.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyphen==0.15.0
PyPika==0.48.9
pyproject_hooks==1.0.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-pptx==0.6.23
python-slugify==8.0.4
python-socketio==5.11.2
python-socks==2.5.3
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.8.2
rake-nltk==1.0.6
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-cache==1.2.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.2
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
selenium==4.21.0
sentence-transformers==2.6.1
sentencepiece==0.2.0
shapely==2.0.4
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
slack_bolt==1.19.0
slack_sdk==3.30.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.36
sqlparse==0.5.1
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
starlette==0.45.3
streamlit==1.36.0
streamlit-aggrid==1.0.5
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-javascript==0.1.5
streamlit-keyup==0.2.4
streamlit-pills==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
striprtf==0.0.26
sympy==1.12
tenacity==8.5.0
text-unidecode==1.3
textstat==0.7.3
threadpoolctl==3.4.0
tifffile==2024.5.10
tiktoken==0.8.0
timm==0.9.16
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.2
transformers==4.39.3
trio==0.25.1
trio-websocket==0.11.1
typer==0.12.3
types-requests==2.32.0.20240622
typing-inspect==0.8.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
url-normalize==2.2.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
validators==0.35.0
watchfiles==1.0.5
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.17.1
zipp==3.22.0
zstandard==0.23.0
